implement rooms page
