+++ themes/hotel-management-system/hotel-management-theme/assets/data/roomTypes.js
@@ -1,0 +1,7 @@
+const RoomTypes = [
+  { id: "1", type: "Luxury" },
+  { id: "2", type: "Family" },
+  { id: "3", type: "Couple" },
+]
+
+export default RoomTypes
