+++ themes/hotel-management-system/hotel-management-theme/src/components/booking/index.js
@@ -1,98 +1,77 @@
-import React from "react"
+import React, { useState } from "react"
 import Grid from "@material-ui/core/Grid"
 import Typography from "@material-ui/core/Typography"
-import Divider from "@material-ui/core/Divider"
-import Button from "@material-ui/core/Button"
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
-import { faCheck, faXmark } from "@fortawesome/free-solid-svg-icons"
+import SelectComponent from "../food/select"
+import BookingCard from "./bookingCard"
 import Styles from "./styles"
+import RoomList from "../../../assets/data/roomList"
+import RoomTypes from "../../../assets/data/roomTypes"
+import {
+  faBuilding,
+  faUser,
+  faDollar,
+  faBed,
+} from "@fortawesome/free-solid-svg-icons"
 
-const BookingCard = ({
-  imgUrl,
-  title,
-  description,
-  roomId,
-  moreInfoList,
-  availability,
-}) => {
-  const classes = Styles({ availability })
+const RoomsPageTemplate = () => {
+  const [type, setType] = useState("")
 
-  const moreInfo = moreInfoList.map(ele => (
-    <Grid item key={ele.id} classes={{ root: classes.detailStyle }}>
-      <Button
-        startIcon={
-          <FontAwesomeIcon icon={ele.icon} className={classes.iconStyle} />
-        }
-        classes={{ root: classes.normalBtnStyle }}
-      >
-        {ele.key}: {ele.value}
-      </Button>
-    </Grid>
-  ))
-  return (
-    <Grid container direction="column" classes={{ root: classes.container }}>
-      <Grid item>
-        <img src={imgUrl} className={classes.imgStyle} />
-      </Grid>
-      <Grid container direction="column" classes={{ root: classes.wrapper }}>
-        <Grid item>
-          <Typography align="center">{title}</Typography>
-        </Grid>
-        <Grid item>
-          <Divider classes={{ root: classes.divideStyle }} />
-        </Grid>
-        <Grid item classes={{ root: classes.space }}>
-          <Typography variant="h6">{description}</Typography>
-        </Grid>
-      </Grid>
-      <Grid item>
-        <Grid
-          container
-          direction="column"
-          classes={{ root: classes.fullWidthContainer }}
-        >
-          <Grid item container justifyContent="center">
-            <Typography variant="h4">Room Number: {roomId}</Typography>
+  const handleChange = event => {
+    setType(event.target.value)
+  }
+
+  const roomItems =
+    type === ""
+      ? RoomList.map(item => (
+          <Grid item key={item.id}>
+            <BookingCard
+              imgUrl={item.imgUrl}
+              title={item.title}
+              description={item.description}
+              roomId={item.roomId}
+              moreInfoList={item.moreInfoList}
+              availability={item.availability}
+            />
           </Grid>
-        </Grid>
+        ))
+      : RoomList.filter(ele => ele.type === type).map(item => (
+          <Grid item key={item.id}>
+            <BookingCard
+              imgUrl={item.imgUrl}
+              title={item.title}
+              description={item.description}
+              roomId={item.roomId}
+              moreInfoList={item.moreInfoList}
+              availability={item.availability}
+            />
+          </Grid>
+        ))
+  const classes = Styles()
+  return (
+    <Grid container direction="column" alignItems="center">
+      <Grid item classes={{ root: classes.headerStyle }}>
+        <Typography variant="h3" align="center">
+          Rooms at<span className={classes.decoratedHeader}> Cloud Hotel</span>
+        </Typography>
+      </Grid>
+      <Grid item classes={{ root: classes.selectStyle }}>
+        <SelectComponent
+          title="Type"
+          value={type}
+          onChangeHandler={handleChange}
+          dataList={RoomTypes}
+        />
       </Grid>
       <Grid
         item
         container
-        direction="column"
-        classes={{ root: classes.wrapper }}
+        direction="row"
+        classes={{ root: classes.itemContainer }}
       >
-        <Grid item container direction="row">
-          {moreInfo}
-        </Grid>
-        <Grid
-          item
-          container
-          direction="row"
-          classes={{ root: classes.space }}
-          justifyContent="space-around"
-        >
-          <Grid item>
-            <Button
-              startIcon={
-                availability ? (
-                  <FontAwesomeIcon icon={faCheck} />
-                ) : (
-                  <FontAwesomeIcon icon={faXmark} />
-                )
-              }
-              classes={{ root: classes.avaiableBtnStyle }}
-            >
-              {availability ? "Available" : "Unavailable"}
-            </Button>
-          </Grid>
-          <Grid item>
-            <Button classes={{ root: classes.bookBtnStyle }}>Book Now</Button>
-          </Grid>
-        </Grid>
+        {roomItems}
       </Grid>
     </Grid>
   )
 }
 
-export default BookingCard
+export default RoomsPageTemplate
